requirements.txt
django==4.2.30
djangorestframework==3.16.1
gunicorn==23.0.0
openpyxl==3.1.5
psycopg2-binary==2.9.11
python-dotenv==1.2.1
